Create entity package
