+++ src/main/java/com/wellseecoding/server/entity/post/PostRepository.java
@@ -1,4 +1,4 @@
-package com.wellseecoding.server.post;
+package com.wellseecoding.server.entity.post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
