+++ src/main/java/com/wellseecoding/server/entity/sns/SnsInfoKey.java
@@ -1,4 +1,4 @@
-package com.wellseecoding.server.user.sns;
+package com.wellseecoding.server.entity.sns;
 
 import lombok.*;
